+++ RELEASES.md
@@ -1,6 +1,8 @@
-# Version 0.1.1 (2021-03-06)
+# Version 0.1.1 (2021-03-11)
 
   * Reliably build documentation at <https://doc.qu1x.dev/trackball>.
+  * Clamp cursor/finger position between zero and maximum position.
+  * Add identical [C11 implementation](c11).
 
 # Version 0.1.0 (2021-03-06)